+++ src/RadixDlt.NetworkGateway.PostgresIntegration/GatewayModelExtensions.cs
@@ -85,6 +85,16 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration;
 
 internal static class GatewayModelExtensions
 {
+    public static GatewayModel.AccountResourcePreferencesResponseItem.ResourcePreferenceRuleEnum ToGatewayModel(this AccountResourcePreferenceRule input)
+    {
+        return input switch
+        {
+            AccountResourcePreferenceRule.Allowed => GatewayModel.AccountResourcePreferencesResponseItem.ResourcePreferenceRuleEnum.Allowed,
+            AccountResourcePreferenceRule.Disallowed => GatewayModel.AccountResourcePreferencesResponseItem.ResourcePreferenceRuleEnum.Disallowed,
+            _ => throw new ArgumentOutOfRangeException(nameof(input), input, null),
+        };
+    }
+
     public static GatewayModel.ManifestClass ToGatewayModel(this LedgerTransactionManifestClass input)
     {
         return input switch